refactor: 접근 제한자 수정

+++ src/main/java/chess/domain/piece/Bishop.java
@@ -7,7 +7,7 @@ public class Bishop extends Piece {
 
     private static final String BLACK_SYMBOL = "B";
     private static final String WHITE_SYMBOL = "b";
-    private static final int BISHOP_SCORE = 3;
+    private static final double BISHOP_SCORE = 3.0;
 
     public Bishop(final Team team) {
         super(team);
