+++ src/main/java/chess/domain/piece/Rook.java
@@ -7,7 +7,7 @@ public class Rook extends Piece {
 
     private static final String BLACK_SYMBOL = "R";
     private static final String WHITE_SYMBOL = "r";
-    private static final int ROOK_SCORE = 5;
+    private static final double ROOK_SCORE = 5;
 
     public Rook(final Team team) {
         super(team);
